io: i2c: implement a software I2C bus
Add a software implementation of the I2C protocol. The only
dependency of this module is a fully functional GPIO chip. Even though
an external GPIO expander can technically (i.e. theoretically) be used
as GPIO chip, it is highly recommended (read: required) to use the
GPIO chip of the SoC. Multiple threads can use the algorithm at the
same time, as both transfer methods, `setFrequency()' and `test()' are
protected by a mutex.

An application can test the current state of a bus by calling
`SoftareI2CAlgorithm::test()'. This method will fully test both the clock
and data lines in responsiveness.

This implementation uses a variadic frequency; meaning that the frequency
of the bus can be configured using `I2CAlgorithm::setFrequency()'.

Signed-off-by: Michel Megens <dev@bietje.net>