io: i2c: add a busy check to software I2C transfers
Implement a method to check if the bus is busy before initiating a
master transfer. This is done by sampling the SDA and SCL lines twice
with a 10ms period between. If either is pulled low, the transfer is
cancelled with the `ETRYAGAIN' return code. This means that the I2C bus
will attempt a retransmission, for as long retransmissions are still
available.

Signed-off-by: Michel Megens <dev@bietje.net>