kernel: uniquelock: implement uniquelock
Implement `UniqueLock<T>'. Like `ScopedLock', this class implements a
managed lock using RAII. The template parameter is used to define the
implementation of the actual lock it wraps. This means that `ScopedLock'
can be defined as a function of UniqueLock:

	ScopedLock = UniqueLock<Lock>

In the future the definition of ScopedLock might be removed and changed
to the using statement given above.

Signed-off-by: Michel Megens <dev@bietje.net>