io: i2c: implement a hardware I2C algorithm
Implement a hardware I2C algorithm based on the I2C hardware abstraction
layer. The hardware abstraction layer for I2C is implemented by the
platform library.  The hardware I2C algorithm implements the I2C algorithm
on the byte level. The bitwise pin operations are an impelentation detail
which is usually implemented in hardware. Failed transmissions are
automatically retried by either the algorithm or the I2C bus controller.
The HAL implementation does not need to worry about retransmissions.

The HAL implementation can either be implemented in realtime, or by
buffering the commands. The function `i2c_write_buffers()' is required to
complete the I2C transmission. This funcion can be empty if the HAL uses
a realtime implementation.

Signed-off-by: Michel Megens <dev@bietje.net>