io: i2c: improve hardware I2C logging
Improve the logging output of the hardware I2C implementation. The logging
instance is initialized with the prefix "hw-i2c" in the `HardwareI2CAlgorithm'
constructor. Further more, the single message and multi message transfer
methods log messages on the same errors. Errors are logged to the stdout stream
when:

	* the I2C bus is busy;
	* the hardware reports an error.

The busy error will only be logged in debug builds. Hardware errors are always
logged on `stdout'.

Signed-off-by: Michel Megens <dev@bietje.net>