io: i2c: implement hardware I2C as an abstract class
The current implementation of the hardware I2C algorithm uses an API
called the "I2C HAL". This HAL is a C header which defines the I2C
operations on a byte level. The platform libraries are responsible for
the implementation of the I2C HAL. This patch removes the traditional
HAL header, and implements the same functionality in the hardware I2C
class itself. This means that the current `HardwareI2CAlgorithm' class has
been marked as abstract, and all I2C HAL functions are implemented as
abstract methods. The platform libraries are expected to subclass this class
and implement these methods.

This patch is implemented to allow hardware abstraction layers to define
multiple hardware I2C algorithms.

Further more, this patch reduces  chances of cryptic linker messages when using
the hardware I2C algorithm while the I2C HAL isn't implemented by the platform
libraries. The new situation disables the use of the hardware I2C entirely if
the correct methods are not implemented and the compiler instead of the linker
will throw these errors.

Signed-off-by: Michel Megens <dev@bietje.net>