io: i2c: update transfer method declaration
Update the declaration of the transfer method. The new declaration uses a
vector of I2C messages, instead of a vector of pointers to I2C messages. This
implementation improves the safety of I2C driver code, by reducing the usage
of raw pointers.

Performance of I2C drivers doesn't need to suffer (due to copy operations when
inserting new elements). The I2C vector can be created and/or updated using move
semantics through the `Vector::pushback()' method. Similarly, the results of an
I2C operation can be retrieved using move sematics and the `Vector::back()'
method.

Signed-off-by: Michel Megens <dev@bietje.net>