kernel: fsm: implement a finite state machine
A finite state machine (FSM) is a mathematical model of computation. It
is an abstract machine that can only be in one of a finite number of
states. A transition from one state to the next can either be initiated
by the state execution itself, or by external input. A FSM is defined
by defined by a quintuple (E, S, s0, d, F), where:

	* E is the input alphabet, a finite non-empty set of symbols
	* S is a finite non-empty set of states
	* s0 is the initial state (member of S)
	* d is the transition function: d : S x E -> S
	* F is a finite, possibly empty, set of final states (subset of S)

Programs can define the transition function using the methods `raise()',
`raiseFromIrq()' and `transition()'. The latter can be used to
transition from one state to another (based on an input symbol) from within
a state. The former two can be used by external components to influence the
FSM.

The FSM implementation supports both DFA's and NDFA's. The determinism of a
FSM can be checked using the `deterministic()' method.

While the `detail::FsmBase' class provides a solid base for FSM's, it needs
specialized methods in order to execute the FSM. An asynchronous
implementation has been provided in the `AsyncFsm' class. A synchronous
version is implement in the class `FSM'.

Signed-off-by: Michel Megens <michel@michelmegens.net>