kernel: dqp: implement a dispatch queue
A dispatch queue is queue to which an application can submit executable
objects. The configuration of a dispatch queue is determined by the
policy of the dispatch queue. The lwIoT framework supplies two default
policies: the multi-threading policy and the single threaded policy.

Work items (i.e. an executable object) is scheduled on the dispatch
queue in an asynchronous manner. This means that the executable
object will be executed at some point in the future, while the context
that submitted the work continues execution.

The dispatch queue will execute work in a thread. This means that any
method or function can in theory be called. Blocking methods should always
be avoided, since all work items on the dispatch queue share the same
executing thread. Blocking this thread results in blocking the entire
dispatch queue.

Signed-off-by: Michel Megens <michel@michelmegens.net>