kernel: fsm: refactor the transition function implementation
Implement the transition function using a single map object. Previously
the transition information was spread across different states and stored
in dynamic array types. While this makes sense from an OO-programming
standpoint, it is not very efficient. The key of this map object consists
partially of the state ID and part of the event ID. To facilitate this
the creation of these keys, a union type has been created: the `SttIndex'.

As a result, transitions are no longer added to a state, but to the
state machine.  While this might seem counter-intuitive, it makes a lot
of sense from a computational point of view. The complexity of a
transition has been improved from O(n) to O(log n). The code complexity
for the deterministic FSM check is also reduced.

Transitions can be added to an FSM using the `FsmBase::addTransition()'
method. This method implements perfect forwarding to the transition
constructor.

Signed-off-by: Michel Megens <michel@michelmegens.net>